% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb_analysis.R
\name{extract_verb_details}
\alias{extract_verb_details}
\title{Extract details from various dplyr verbs}
\usage{
extract_verb_details(call, max_show = 5)
}
\description{
Extract details from various dplyr verbs
}
\keyword{internal}
