% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipe_colors}
\alias{pipe_colors}
\title{Get the color scheme used for different verb types}
\usage{
pipe_colors()
}
\value{
A named list of hex color codes
}
\description{
Returns a named list of colors used in pipeline visualizations.
}
\examples{
\dontrun{
colors <- pipe_colors()
colors$filter  # "#FFF4E6"
}
}
