% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{pipe_graph}
\alias{pipe_graph}
\title{Build a graph representation from a pipeline with verb metadata}
\usage{
pipe_graph(expr)
}
\arguments{
\item{expr}{Pipeline expression}
}
\value{
list(nodes=..., edges=...)
}
\description{
Build a graph representation from a pipeline with verb metadata
}
