% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipe_summary}
\alias{pipe_summary}
\title{Print a text summary of a pipeline}
\usage{
pipe_summary(expr)
}
\arguments{
\item{expr}{A pipeline expression}
}
\value{
Pipeline parts (invisibly)
}
\description{
Displays a human-readable summary of the pipeline structure without
rendering a visualization.
}
\examples{
\dontrun{
pipe_summary(quote(
  mtcars \%>\%
    filter(cyl > 4) \%>\%
    mutate(efficiency = mpg / hp)
))
}
}
