% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_viz.R
\name{pipe_vizr_file}
\alias{pipe_vizr_file}
\title{Visualize dplyr pipelines from a file}
\usage{
pipe_vizr_file(
  file,
  combine = TRUE,
  direction = "LR",
  theme = "default",
  show_expressions = TRUE,
  fixed_width = TRUE
)
}
\arguments{
\item{file}{Path to .R, .Rmd, or .Qmd file}

\item{combine}{If TRUE, show all pipelines in one diagram. If FALSE, 
create separate diagrams for each pipeline.}

\item{direction}{Graph layout direction}

\item{theme}{Visual theme}

\item{show_expressions}{Show RHS of assignments}

\item{fixed_width}{Make all nodes same width}
}
\value{
A visualization or list of visualizations
}
\description{
Extract and visualize all dplyr pipelines from an R script, R Markdown, 
or Quarto file.
}
\examples{
\dontrun{
# Visualize all pipelines in a script
pipe_vizr_file("my_analysis.R")

# Separate diagram for each pipeline
pipe_vizr_file("my_analysis.R", combine = FALSE)

# From R Markdown
pipe_vizr_file("report.Rmd")
}

}
